# Allow headers to match on ReferenceEquals before OrdinalIgnoreCase