--- a/src/Hosting/Hosting/src/Internal/WebHost.cs
+++ b/src/Hosting/Hosting/src/Internal/WebHost.cs
@@ -22,6 +22,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.StackTrace.Sources;
+using Microsoft.Net.Http.Headers;
 
 namespace Microsoft.AspNetCore.Hosting.Internal
 {
@@ -276,7 +277,7 @@ namespace Microsoft.AspNetCore.Hosting.Internal
                 return context =>
                 {
                     context.Response.StatusCode = 500;
-                    context.Response.Headers["Cache-Control"] = "no-cache";
+                    context.Response.Headers[HeaderNames.CacheControl] = "no-cache";
                     return errorPage.ExecuteAsync(context);
                 };
             }