--- a/src/Http/Http.Abstractions/ref/Microsoft.AspNetCore.Http.Abstractions.csproj
+++ b/src/Http/Http.Abstractions/ref/Microsoft.AspNetCore.Http.Abstractions.csproj
@@ -7,6 +7,7 @@
     <Compile Include="Microsoft.AspNetCore.Http.Abstractions.netcoreapp3.0.cs" />
     <Reference Include="Microsoft.AspNetCore.Http.Features"  />
     <Reference Include="Microsoft.Extensions.ActivatorUtilities.Sources"  />
+    <Reference Include="Microsoft.Net.Http.Headers"  />
     <Reference Include="System.Text.Encodings.Web"  />
   </ItemGroup>
 </Project>