--- a/src/Middleware/Session/ref/Microsoft.AspNetCore.Session.csproj
+++ b/src/Middleware/Session/ref/Microsoft.AspNetCore.Session.csproj
@@ -10,5 +10,6 @@
     <Reference Include="Microsoft.Extensions.Caching.Abstractions"  />
     <Reference Include="Microsoft.Extensions.Logging.Abstractions"  />
     <Reference Include="Microsoft.Extensions.Options"  />
+    <Reference Include="Microsoft.Net.Http.Headers"  />
   </ItemGroup>
 </Project>