… src/Http/Headers/src/HeaderNames.cs
… src/Http/Http.Abstractions/src/Microsoft.AspNetCore.Http.Abstractions.csproj
… src/Http/Http/src/Internal/DefaultHttpRequest.cs
… src/Http/Http/src/Internal/FormFile.cs
… src/Http/Routing/src/Matching/HttpMethodMatcherPolicy.cs
… src/Http/WebUtilities/src/MultipartSection.cs
… src/Middleware/Session/src/Microsoft.AspNetCore.Session.csproj
… src/Middleware/Session/src/SessionMiddleware.cs
… src/Middleware/WebSockets/ref/Microsoft.AspNetCore.WebSockets.netcoreapp3.0.cs
… src/Middleware/WebSockets/src/Internal/Constants.cs
… src/Middleware/WebSockets/src/Internal/HandshakeHelpers.cs
… src/Middleware/WebSockets/src/WebSocketMiddleware.cs
… src/Middleware/WebSockets/test/UnitTests/WebSocketMiddlewareTests.cs
… src/Mvc/test/Mvc.FunctionalTests/RequestServicesTestBase.cs
… src/Mvc/test/WebSites/BasicWebSite/RequestIdMiddleware.cs
… src/Security/Authentication/Cookies/src/CookieAuthenticationEvents.cs
… src/Security/Authentication/JwtBearer/src/JwtBearerHandler.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpHeaders.Generated.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpHeaders.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpProtocol.FeatureCollection.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpProtocol.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpRequestHeaders.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpResponseHeaders.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpResponseTrailers.cs
… src/Servers/Kestrel/perf/Kestrel.Performance/ResponseHeaderCollectionBenchmark.cs
… src/Servers/Kestrel/perf/Kestrel.Performance/ResponseHeadersWritingBenchmark.cs
… src/Servers/Kestrel/shared/KnownHeaders.cs
… src/Servers/Kestrel/test/InMemory.FunctionalTests/Http2/Http2StreamTests.cs
… src/SignalR/common/Http.Connections/src/Internal/Transports/ServerSentEventsTransport.cs