Provide a performance advantage to using the `Microsoft.Net.Http.Headers.HeaderNames` constants.

Change `HeaderNames` to static properties rather than string constants; so reference equality works; prior to falling back to an OrdinalIgnoreCase characters comparison. (string constants embed directly in the calling assembly so don't end up being reference equal)

Use these properties in Kestrel header methods rather than its own string constants so the reference equality works there.

ResponseHeaderCollectionBenchmark (Common is the main change +18%)
```diff
      Method |                 Type |        Mean |         Op/s | Allocated |
 ----------- |--------------------- |------------:|-------------:|----------:|
- SetHeaders |               Common |   889.92 ns |  1,123,698.8 |       0 B |
+ SetHeaders |               Common |   751.84 ns |  1,330,077.7 |       0 B |
- SetHeaders |  ContentLengthString |    51.64 ns | 19,363,867.1 |       0 B |
+ SetHeaders |  ContentLengthString |    51.90 ns | 19,268,524.6 |       0 B |
- SetHeaders |            Plaintext |    55.45 ns | 18,034,860.0 |       0 B |
+ SetHeaders |            Plaintext |    51.40 ns | 19,455,753.2 |       0 B |
- SetHeaders |              Unknown | 1,539.01 ns |    649,770.4 |       0 B |
+ SetHeaders |              Unknown | 1,589.23 ns |    629,237.4 |       0 B |
```

The inlined string property becomes a constant mov at Jit time e.g.
```csharp
if (ReferenceEquals(HeaderNames.Host, key))
```
becomes
```asm
G_M13073_IG06:
       mov      rdx, 0x15B8E12EBA8        ; HeaderNames:get_Host():ref
       cmp      gword ptr [rdx], rdi      ; ReferenceEquals
       jne      SHORT G_M13073_IG07
```

For plaintext with the default builder (including host filtering)

Before
![image](https://user-images.githubusercontent.com/1142958/56084029-090a9480-5e25-11e9-9742-04bd80ed7043.png)

After
![image](https://user-images.githubusercontent.com/1142958/56084061-7f0efb80-5e25-11e9-9651-0418e591bed0.png)

----
Background

Noticed this while looking at something else https://github.com/aspnet/AspNetCore/pull/9359

![image](https://user-images.githubusercontent.com/1142958/56074785-2e10f000-5db0-11e9-89a2-c2f3e0cc7ba1.png)

If you use the `HeaderNames` class they don't match with reference equality e.g.
```csharp
context.Request.Headers[HeaderNames.Host]
```
as they are defined in a different assembly to the string constants used in Kestrel; so it moves to an OrdinalIgnoreCase compare. Using the same `HeaderNames` header names in Kestrel means it will first match using [ReferenceEquals in coreclr](https://github.com/dotnet/coreclr/blob/0fed62e68d1f8f8a756ec8ba5348bfde61057666/src/System.Private.CoreLib/shared/System/String.Comparison.cs#L638-L644) 
```
public bool Equals(string? value, StringComparison comparisonType)
{
    if (object.ReferenceEquals(this, value))
    {
        CheckStringComparison(comparisonType);
        return true;
    }
```
and won't need to fallback to the `OrdinalIgnoreCase` when `HeaderNames` is used e.g when setting
```csharp
response.ContentType = "text/plain";
```
as [DefaultHttpResponse](https://github.com/aspnet/AspNetCore/blob/e04c79bd3dfc0634af1ac6b578124630a5be247d/src/Http/Http/src/Internal/DefaultHttpResponse.cs#L89-L99) also uses these constants
```csharp
set
{
    if (string.IsNullOrEmpty(value))
    {
        HttpResponseFeature.Headers.Remove(HeaderNames.ContentType);
    }
    else
    {
        HttpResponseFeature.Headers[HeaderNames.ContentType] = value;
    }
}
```