@benaadams 2019-04-13
Need to investigate more as they are consts in the ref assembly so it just might embed them and not work.