@benaadams 2019-04-13
Works now, though the string properties don't inline https://github.com/dotnet/coreclr/issues/23969; however reference equality works without needing to fall back to the OrdinalIgnoreCase compare.