@benaadams 2019-04-14
ResponseHeaderCollectionBenchmark
```diff
      Method |                 Type |        Mean |         Op/s | Allocated |
 ----------- |--------------------- |------------:|-------------:|----------:|
- SetHeaders |               Common |   889.92 ns |  1,123,698.8 |       0 B |
+ SetHeaders |               Common |   751.84 ns |  1,330,077.7 |       0 B |
- SetHeaders |  ContentLengthString |    51.64 ns | 19,363,867.1 |       0 B |
+ SetHeaders |  ContentLengthString |    51.90 ns | 19,268,524.6 |       0 B |
- SetHeaders |            Plaintext |    55.45 ns | 18,034,860.0 |       0 B |
+ SetHeaders |            Plaintext |    51.40 ns | 19,455,753.2 |       0 B |
- SetHeaders |              Unknown | 1,539.01 ns |    649,770.4 |       0 B |
+ SetHeaders |              Unknown | 1,589.23 ns |    629,237.4 |       0 B |
```