@benaadams 2019-04-15
I may have sent these changes from the future?
```
F:\workspace\_work\1\s\build\tasks\RepoTasks.csproj :
 error MSB4236: The SDK 'Microsoft.NET.Sdk' specified could not be found.
```