@Eilon 2019-04-16
I think @javiercn was looking at YARN reliability.