@halter73 2019-04-16
> Can you restart that job? 😉 I only have the power to restart everything, not an individual item.

Generally I wait until all the builds are complete before triaging and restarting the failed build. I'm not sure if you can restart just the failed build before then. I'm setting a reminder to requeue once this is finished.