@halter73 2019-04-16
> Yeah, but this PR needs to be rebased on top of master.

Doesn't our PR validation automatically rebase on the target branch?