@javiercn 2019-04-16
> Doesn't our PR validation automatically rebase on the target branch?

It doesn't AFAIK. It simply checks out the repo/branch and builds it, you still have to deal with merges/rebases.