@halter73 2019-04-17
Looks good to go. I'll merge this tomorrow so I remember to make the breaking change announcement at the same time.