@anurse 2019-04-17
> It doesn't AFAIK. It simply checks out the repo/branch and builds it, you still have to deal with merges/rebases.

It doesn't rebase, but it **does** use the "merge ref" from GitHub (`refs/pull/9341/merge` for example) which contains the **merge** of this PR and `master`. Though, I don't know when GitHub updates this ref (do they do it for every `master` commit or just for every commit to the PR).