@anurse 2019-04-17
At a minimum, it's safe to assume the build is incorporating all changes from the most recent commit to the PR branch **and** all changes that were in `master` **at the time the most recent PR commit was made**